fix(func): remove check link
related #150
