+++ backend/server/calendars_bookings_get.py
@@ -3,14 +3,17 @@ from flask.json import jsonify
 from tools.build_response import build_response
 from tools.for_db.work_with_slots import get_slots_by_admin_id_and_booking_id, BookingNotFound
 from tools.for_db.work_with_booking_info import get_booking_info
-from tools.for_db.work_with_links import check_link, LinkHasExpired, LinkNotFound
+from tools.for_db.work_with_links import get_link
+from tools.check_link import check_link, LinkHasExpired, LinkNotFound
+
 calendars_bookings_get = Blueprint('calendars_bookings_get', __name__)
 
 
 @calendars_bookings_get.route('/calendars/<link_id>/bookings/<booking_id>', methods=['GET'])
 def get_calendars_bookings(link_id, booking_id):
     try:
-        link = check_link(link_id)
+        link = get_link(link_id)
+        check_link(link)
         slots = get_slots_by_admin_id_and_booking_id(link.admin_id, booking_id)
         booking_info = get_booking_info(booking_id)
     except BookingNotFound as e:
